requirements.txt
Pillow==10.1.0
Pillow==10.2.0
protobuf==4.25.2
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.31.0
